[integration tests] Move oscore tests in a dedicated package
